Refresh 3

+++ lib/src/guMtxF2L.c
@@ -2,10 +2,10 @@
 
 void guMtxF2L(float mf[4][4], Mtx *m) {
     int r, c;
-    long tmp1;
-    long tmp2;
-    long *m1 = &m->m[0][0];
-    long *m2 = &m->m[2][0];
+    s32 tmp1;
+    s32 tmp2;
+    s32 *m1 = &m->m[0][0];
+    s32 *m2 = &m->m[2][0];
     for (r = 0; r < 4; r++) {
         for (c = 0; c < 2; c++) {
             tmp1 = mf[r][2 * c] * 65536.0f;
