+++ src/goddard/old_menu.c
@@ -1,5 +1,7 @@
 #include <ultra64.h>
 #include <macros.h>
+#include <stdio.h>
+
 #include "gd_types.h"
 #include "old_menu.h"
 #include "objects.h"
@@ -28,7 +30,7 @@ static struct ObjGadget *sCurGadgetPtr;
 void func_8018BCB8(struct ObjGadget *);
 
 /* 239EC0 -> 239F78 */
-void get_objvalue(union ObjVarVal *dst, enum ValPtrType type, void *base, s32 offset) {
+void get_objvalue(union ObjVarVal *dst, enum ValPtrType type, void *base, size_t offset) {
     union ObjVarVal *objAddr = (void *) ((u8 *) base + offset);
 
     switch (type) {
@@ -124,7 +126,7 @@ struct ObjGadget *make_gadget(UNUSED s32 a0, s32 a1) {
 }
 
 /* 23A32C -> 23A3E4 */
-void set_objvalue(union ObjVarVal *src, enum ValPtrType type, void *base, s32 offset) {
+void set_objvalue(union ObjVarVal *src, enum ValPtrType type, void *base, size_t offset) {
     union ObjVarVal *dst = (void *) ((u8 *) base + offset);
     switch (type) {
         case OBJ_VALUE_INT:
